requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
PyMySQL==1.1.0
python-dateutil==2.8.2
bcrypt==4.0.1
SQLAlchemy-Utils==0.41.1
mysql-connector-python==8.2.0
psutil==5.9.6
pypinyin==0.51.0